requirements.txt
streamlit
pandas
numpy
easyocr
pillow
rapidfuzz
pytesseract
matplotlib
sqlite-utils

